feat: Change target link libraries based on type
If we want to add flags to a target that's an executable, then we
can safely use the PRIVATE attribute when linking in gcov.
For other target types, such as STATIC or SHARED libraries, we
need to make sure that gcov is linked both during building of the
library and other targets dependent on the library.

In our specific use case we only want the project library being
built to have been instrumented. We don't want to instrument the
GoogleTest executable because we don't care to exercise their code.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>