fix: Ensure header files are covered
Some header files include functions and templates that are not present
when instrumenting the source files in libraries.
In order to make sure those are covered when included in a GoogleTest
executable when need to pass the '--coverage' flag to any target that
is linking againt the instrumented library.

An alternative way of handling this is to ensure all functions are
included in source files, but this does not work for templates without
doing partial specialization in the source file, which can become
increasingly complex to manage.

In addition, we need to ensure that the GoogleTest coverage is not
reported. Unfortunately, it will still get instrumented.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>