fix: Enable gcovr summary print
Enabling the GCOVR summary to be printed to the terminal when completed.
NOTE: I do not care for this implementation, but it does give
flexibility for printing additional information to the terminal.
We are using this for GitLab coverage results.
Perhaps the better way to do this is via the cmake command line so we
can optionally define these arguments to gcovr.

Signed-off-by: Ryan Sherlock <ryan.m.sherlock@gmail.com>